rename: CommentRepository 패키지 추가 (#308)
