+++ src/main/java/com/listywave/list/application/service/ListService.java
@@ -37,8 +37,8 @@ import com.listywave.list.application.dto.response.ListTrandingResponse;
 import com.listywave.list.presentation.dto.request.ItemCreateRequest;
 import com.listywave.list.presentation.dto.request.ListCreateRequest;
 import com.listywave.list.presentation.dto.request.ListUpdateRequest;
-import com.listywave.list.repository.CommentRepository;
 import com.listywave.list.repository.ItemRepository;
+import com.listywave.list.repository.comment.CommentRepository;
 import com.listywave.list.repository.label.LabelRepository;
 import com.listywave.list.repository.list.ListRepository;
 import com.listywave.list.repository.reply.ReplyRepository;
